backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
psycopg2-binary
sqlalchemy
pandas
python-dotenv
redis
requests
